Simplify Joshua coverage regression tests
